Move functions tests (#7318)
